# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ProConnect professional networking and remote jobs platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profiles/{profileId}/education:
    get:
      operationId: listEducation
      tags: [education]
      summary: List education entries for a profile
      parameters:
        - name: profileId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of education entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Education"
    post:
      operationId: createEducation
      tags: [education]
      summary: Add education entry
      parameters:
        - name: profileId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEducationBody"
      responses:
        "201":
          description: Education created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Education"
  /profiles/{profileId}/education/{id}:
    put:
      operationId: updateEducation
      tags: [education]
      summary: Update education entry
      parameters:
        - name: profileId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEducationBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Education"
    delete:
      operationId: deleteEducation
      tags: [education]
      summary: Delete education entry
      parameters:
        - name: profileId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /profiles/{profileId}/experience:
    get:
      operationId: listExperience
      tags: [experience]
      summary: List experience entries for a profile
      parameters:
        - name: profileId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of experience entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Experience"
    post:
      operationId: createExperience
      tags: [experience]
      summary: Add experience entry
      parameters:
        - name: profileId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExperienceBody"
      responses:
        "201":
          description: Experience created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Experience"
  /profiles/{profileId}/experience/{id}:
    put:
      operationId: updateExperience
      tags: [experience]
      summary: Update experience entry
      parameters:
        - name: profileId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExperienceBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Experience"
    delete:
      operationId: deleteExperience
      tags: [experience]
      summary: Delete experience entry
      parameters:
        - name: profileId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /jobs/{jobId}/applications:
    get:
      operationId: listApplications
      tags: [applications]
      summary: List applications for a job
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Application"
    post:
      operationId: applyToJob
      tags: [applications]
      summary: Apply to a job
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApplicationBody"
      responses:
        "201":
          description: Application submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /profiles/{profileId}/applications:
    get:
      operationId: listProfileApplications
      tags: [applications]
      summary: List job applications for a profile
      parameters:
        - name: profileId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Application"
...
components:
  schemas:
...
    Profile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        headline:
          type: string
        bio:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        coverUrl:
          type: string
          nullable: true
        website:
          type: string
          nullable: true
        linkedinUrl:
          type: string
          nullable: true
        githubUrl:
          type: string
          nullable: true
        twitterUrl:
          type: string
          nullable: true
        openToWork:
          type: boolean
        accountType:
          type: string
          nullable: true
        industry:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        customLinks:
          type: array
          items:
            type: object
            properties:
              label:
                type: string
              url:
                type: string
            required: [label, url]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, headline, openToWork, createdAt, updatedAt]
...
    Education:
      type: object
      properties:
        id:
          type: integer
        profileId:
          type: integer
        school:
          type: string
        degree:
          type: string
          nullable: true
        fieldOfStudy:
          type: string
          nullable: true
        startYear:
          type: integer
        endYear:
          type: integer
          nullable: true
        description:
          type: string
          nullable: true
      required: [id, profileId, school, startYear]
    CreateEducationBody:
      type: object
      properties:
        school:
          type: string
        degree:
          type: string
          nullable: true
        fieldOfStudy:
          type: string
          nullable: true
        startYear:
          type: integer
        endYear:
          type: integer
          nullable: true
        description:
          type: string
          nullable: true
      required: [school, startYear]
    Experience:
      type: object
      properties:
        id:
          type: integer
        profileId:
          type: integer
        company:
          type: string
        title:
          type: string
        location:
          type: string
          nullable: true
        remote:
          type: boolean
        startDate:
          type: string
        endDate:
          type: string
          nullable: true
        current:
          type: boolean
        description:
          type: string
          nullable: true
      required: [id, profileId, company, title, remote, startDate, current]
    CreateExperienceBody:
      type: object
      properties:
        company:
          type: string
        title:
          type: string
        location:
          type: string
          nullable: true
        remote:
          type: boolean
        startDate:
          type: string
        endDate:
          type: string
          nullable: true
        current:
          type: boolean
        description:
          type: string
          nullable: true
      required: [company, title, remote, startDate, current]
...
    Job:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        company:
          type: string
        companyLogoUrl:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        description:
          type: string
        category:
          type: string
        experienceLevel:
          type: string
        salaryMin:
          type: integer
          nullable: true
        salaryMax:
          type: integer
          nullable: true
        currency:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        featured:
          type: boolean
        applicationCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, title, company, description, category, experienceLevel, tags, featured, applicationCount, createdAt, updatedAt]
...
    Application:
      type: object
      properties:
        id:
          type: integer
        jobId:
          type: integer
        profileId:
          type: integer
        coverLetter:
          type: string
          nullable: true
        status:
          type: string
        appliedAt:
          type: string
        consentToShare: { type: boolean }
        consentScope: { type: array, items: { type: string } }
        consentedAt: { type: string, format: date-time, nullable: true }
        profile:
          $ref: "#/components/schemas/Profile"
        job:
          $ref: "#/components/schemas/Job"
      required: [id, jobId, profileId, status, appliedAt, consentToShare, consentScope, profile, job]
    CreateApplicationBody:
      type: object
      required: [consentToShare]
      properties:
        coverLetter:
          type: string
          nullable: true
        consentToShare:
          type: boolean
          description: Consent to share identity, contact, profile, and cover letter with the job company.
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: >
        HMAC-signed token in the format "profileId:hmac-sha256(profileId, SESSION_SECRET)".
        Returned from /auth/login and /auth/register as authToken.