# lib/api-zod/src/generated/types/createApplicationBody.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * ProConnect professional networking and remote jobs platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateApplicationBody {
  coverLetter?: string | null;
  /** Consent to share identity, contact, profile, and cover letter with the job company. */
  consentToShare: boolean;
}